+++ src/components/projects/index.js
@@ -1,3 +1,5 @@
 import "./projects.css";
 
+export { projectsState } from "./state.js";
+export { projectsMethods } from "./methods.js";
 export { projectsViewMethods } from "./view.js";